refactor sidebar into pathless layout
