+++ apps/webapp/src/components/ChatView.tsx
@@ -1,10 +1,8 @@
 import { useEffect, useLayoutEffect, useMemo, useRef, useState } from "react";
-import { AppSidebar } from "@/components/app-sidebar";
 import { HyperwaveLogoHorizontal, HyperwaveLogoVertical } from "@/components/logo";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { Textarea } from "@/components/ui/textarea";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { cn } from "@/lib/utils";
@@ -180,10 +178,7 @@ export function ChatView({
   };
 
   return (
-    <SidebarProvider>
-      <AppSidebar />
-      <SidebarInset>
-        <div className="relative flex flex-col h-full">
+    <div className="relative flex flex-col h-full">
           <ThreadHeader threadId={threadId} />
           <main
             ref={scrollRef}
@@ -334,7 +329,5 @@ export function ChatView({
             </div>
           </form>
         </div>
-      </SidebarInset>
-    </SidebarProvider>
   );
 }
