+++ apps/webapp/src/routes/_chatLayout.tsx
@@ -1,0 +1,18 @@
+import { Outlet, createFileRoute } from "@tanstack/react-router";
+import { AppSidebar } from "@/components/app-sidebar";
+import { SidebarProvider, SidebarInset } from "@/components/ui/sidebar";
+
+export const Route = createFileRoute("/_chatLayout")({
+  component: ChatLayout,
+});
+
+function ChatLayout() {
+  return (
+    <SidebarProvider>
+      <AppSidebar />
+      <SidebarInset>
+        <Outlet />
+      </SidebarInset>
+    </SidebarProvider>
+  );
+}
